Add files via upload
Stealth Web already exists, but we're not going to paste the code/methods for y'all.  To measure value we allowed an attempt for all-domain validation at US Defense Department (included all branches and Space Force)!

It took more than a week of consented info sharing (and more) for a validation message from them that they finally gained alignment in 5 branches... from the depts of hell we operated (and will operate in, when we want/need to).

They didn't describe how, but it was through methods/tech built for battlefield and not for courtrooms.  Therefore, we are able to say we created an exceptionally stealthy environment, or a very good Stealth Web!

Read and decide if this is something you'd like to work on (with us) for a commercially (civilian viable) product.